requirements.txt
backcall==0.1.0
decorator==4.4.0
Django==2.2.1
django-environ==0.4.5
django-filter==2.1.0
djangorestframework==3.9.4
entrypoints==0.3
flake8==3.7.7
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Markdown==3.1.1
mccabe==0.6.1
parso==0.4.0
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
